component seperation
